sample_code/simple_api_test/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.26.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.5
